requirements.txt
# AI Voice Detection - Requirements
# ==================================

# Core ML/DL
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
shap>=0.41.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.10.0
pyAudioAnalysis>=0.3.14
pydub>=0.25.0

# Speech-to-Text
openai-whisper>=20230314
torch>=1.10.0
torchaudio>=0.10.0

# NLP & Text Processing
spacy>=3.4.0
transformers>=4.20.0
nltk>=3.7

# Vector Database & RAG
faiss-cpu>=1.7.0
chromadb>=0.3.0
sentence-transformers>=2.2.0
langchain>=0.0.200

# LLM APIs
openai>=0.27.0
google-generativeai>=0.1.0

# API Framework
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.64.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Data handling
joblib>=1.1.0
